Add alternate method to define values for table cells in enrollment list admin
Rendering a twig template for each cell gets expensive very quickly,
especially when starting up the engine to render each cell separately for the CSV export.
Most cells are plain-text only, which have no need for fancy twig rendering, a
callback which returns the value does the job as well.