Revert "Replace collection of single choice fields with one multiple choice field"
Using one multiple choice field for each authorization type takes too much place
in the admin interface and causes other functionality to be a few screens lower.
This is not desireable, because this other functionality is more important than
authorized groups.

Using a collection of dropdown menus also makes it more clear which groups are
selected, because they are the only ones visible by name instead of having a tiny
checked checkbox in front of the name.

This reverts commit 395963dbea39ee4d440532714d5b8b0230724fb0.