[PrefillUserDataPlugin] Thoroughly document interactions with already filled-in forms
On first sight, this plugin seems to overwrite user data with information of the
logged in admin user when the admin is viewing or editing the submission.
Thanks to some quite complex interations in the symfony form builder,
this does not happen, but this is not obvious at all.
Hopefully the added comments will prevent future panic about accidentally
overwriting user data