Make form EDIT, LIST_ENROLLMENTS and EDIT_ENROLLMENTS inherit permissions logically
It makes not much sense for a group to have EDIT permission, but no EDIT_ENROLLMENTS or LIST_ENROLLMENTS permission
because they can grant these themselves anyways.
Likewise EDIT_ENROLLMENTS not having LIST_ENROLLMENTS permissions is not useful,
because they can not see the enrollments they have permission to edit.