[AdmissionCheckPlugin] Add AdmissionCheckPlugin
The AdmissionCheck plugin allows to control admission by giving the user a QR-code.
When the QR-code is scanned at the entrance (by someone with EDIT_ENROLLMENT permission on the form), a flag is set on the enrollment that it has already been used for entry.
The AdmissionCheckEvent allows other plugins to vote on whether admission should be granted or not.