Mark plugin configuration root as non-virtual and unmapped
Children of virtual (inherit_data) fields do not appear to have errors correctly mapped to them. Instead the errors
are propagated to the root form, where they are displayed.
With large forms as used to configure all plugins, the error message is far away from the spot of the error,
which is bad UX for the user. Since the errors from constraints are worded generically, it may be impossible to
easily determine which field caused an error.