[ShortUrlPlugin] Only create entity when enabled
A ShortUrl entity should only be created when the ShortUrlPlugin
is actually enabled, else a broken ShortUrl is created without a slug.
This ShortUrl without slug violates the NOT NULL constraint on the slug
column, causing the entity manager to fail when flushing changes.

This breaks creating and editing forms, even when the plugin is not
activated, which is very bad.